changed float type to double in constructor

+++ app/src/main/java/com/pizzapp/model/pizza/Crust.java
@@ -9,7 +9,7 @@ public class Crust {
         this.price = 0;
     }
 
-    public Crust(String name, float price) {
+    public Crust(String name, double price) {
         this.name = name;
         this.price = price;
     }
